Hide public impl detail in the plumbing module
