+++ src/derived.rs
@@ -1,15 +1,15 @@
+use crate::plumbing::CycleDetected;
+use crate::plumbing::QueryDescriptor;
+use crate::plumbing::QueryFunction;
+use crate::plumbing::QueryStorageOps;
+use crate::plumbing::UncheckedMutQueryStorageOps;
 use crate::runtime::ChangedAt;
 use crate::runtime::QueryDescriptorSet;
 use crate::runtime::Revision;
 use crate::runtime::Runtime;
 use crate::runtime::RuntimeId;
 use crate::runtime::StampedValue;
-use crate::CycleDetected;
 use crate::Database;
-use crate::QueryDescriptor;
-use crate::QueryFunction;
-use crate::QueryStorageOps;
-use crate::UncheckedMutQueryStorageOps;
 use log::debug;
 use parking_lot::Mutex;
 use parking_lot::{RwLock, RwLockUpgradableReadGuard};
